requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.4
Pillow==10.4.0
python-dateutil==2.9.0
flask-cors==4.0.1 